Set up sidebar for map page. The sidebar should have scrollerable destination card list and a sticky header with trecipe name.
