+++ src/components/CardMenu/CardMenu.tsx
@@ -8,6 +8,7 @@ import { Menu, MenuItem } from "../Menu/Menu";
  */
 export interface CardMenuProps {
   menuItems: MenuItem[];
+  editIconColor?: string;
 }
 
 /**
@@ -41,6 +42,7 @@ export class CardMenu extends React.Component<CardMenuProps, CardMenuState> {
         <Button
           onClick={this.handleClick}
           icon="ellipsis-h"
+          iconColor={this.props.editIconColor}
           fontSize={1.8}
           text=""
         />
