+++ src/components/variables.scss
@@ -25,7 +25,8 @@ $button-font-size: 0.875rem;
 $input-font-size: 0.8rem;
 $default-font-size: 1rem;
 $title-font-size: 2rem;
-$sub-text-font-size: 0.7rem;
+$subtitle-font-size: 1.5rem;
+$subtext-font-size: 0.7rem;
 
 // --- Element Widths ---//
 $tc-min-width: 350px;
