+++ src/pages/Map/map.scss
@@ -1,0 +1,40 @@
+@import "../../components/variables";
+
+.content-wrapper {
+  display: flex;
+  flex-direction: row;
+  overflow: hidden;
+
+  .map-side-bar {
+    flex: 0 0 400px;
+    height: 100vh;
+    padding: 0 0.2rem;
+
+    .trecipe-header {
+      position: sticky;
+      top: 0;
+      display: flex;
+      flex-direction: row;
+      justify-content: space-between;
+      align-items: center;
+      font-size: $subtitle-font-size;
+      font-family: $default-font-family;
+      background: $tc-white;
+      padding: 1rem 0.5rem;
+    }
+
+    .dest-card-list {
+      overflow: auto;
+      max-height: calc(100vh - 4rem);
+      margin: 0;
+      padding: 0;
+    }
+  }
+
+  .map-container {
+    flex: 1 1 100%;
+    background-size: cover;
+    background-image: url(./sampleMap.png);
+    background-position: center center;
+  }
+}
