+++ src/pages/Trecipe/DestinationCard/destinationCard.scss
@@ -35,7 +35,7 @@
     .dest-address {
       margin-top: 0;
       margin-bottom: 0.5rem;
-      font-size: $sub-text-font-size;
+      font-size: $subtext-font-size;
       color: $silver-gray;
     }
